Huge update consists of: - Created React frontend application - Locally deployed backend and launched server of static files - Temporary realized storage of refresh token in local storage - Upgraded CORS configuration + added up filter for headers in response - Realized accounting of visit after main component load. Accounting happens once per 24h. - Realized interceptor of protected requests to insert refresh jwt - Realized receiving of categories from server to frontend
- Realized receiving of categories from server to frontend
- Realized adaptive navbar component
- Realized sidebar navigation for mobile devices
- Realized forming of categories tree in navigational panel

- Started to use redux pattern
- Updated operations with basket on backend part
- Realized redux state for basket
- Realized DTO for basket receiving and mapping
- Realized receiving basket from server + preview in side popup window
- Realized generic component for modal window creation
- Realized modal window to require agreement for basket synchronization with server
- Realized correct saving of basket to server after changes
- Realized saving of basket changes on tab closing or reloading

- Updated categories table. Added up image for category.
- Realized correct adding of image for simple and repeating category
- Realized output of categories tile with images

- Realized own hook for routes setting
- Realized authentication request in routing hook (in other places it didn’t work correctly)
- Realized authentication form for authorization and registration
- Realized shifting between authorization and registration mode in one window
- Realized authorization for with validation of whole fields
- Realized registration form with validation of whole fields and checking username and email for uniqueness
- Realized form for email confirmation in modal window
- Realized component for email confirmation by link and token

- Realized component for errors output from whole redux stores
- Realized block of notifications in toast style
- Realized output of 3 toasts on screen maximally at one moment of time
- Updated generation of default image for user

+++ back-end/src/main/java/gp/wagner/backend/domain/dto/request/crud/BasketRequestDto.java
@@ -20,7 +20,7 @@ public class BasketRequestDto {
 
     // Ассоциативная коллекция - id варианта товара + его количество
     @NotNull
-    private Map<Integer, Integer> productVariantIdAndCount;
+    private Map<Long, Integer> productVariantIdAndCount;
 
     @Nullable
     private Long userId;
